// src/components/payment/enums.ts
// Enum for bank types
export enum BankType {
  PUBLIC_SECTOR = 'PUBLIC_SECTOR',
  PRIVATE_SECTOR = 'PRIVATE_SECTOR',
  FOREIGN = 'FOREIGN',
  SMALL_FINANCE = 'SMALL_FINANCE',
  REGIONAL_COOPERATIVE = 'REGIONAL_COOPERATIVE',
  FINTECH = 'FINTECH'
}

// Enum for fintech categories
export enum FintechCategory {
  DIGITAL_BANKING = 'DIGITAL_BANKING',
  NEOBANK = 'NEOBANK',
  CORPORATE_PAYMENTS = 'CORPORATE_PAYMENTS',
  PAYMENT_CARDS = 'PAYMENT_CARDS'
}